fix: updated structure and paths
