+++ src/components/HorizontalLinearStepper/index.tsx
@@ -1,8 +1,8 @@
 import Step from '@mui/material/Step';
 
 import { getStepContent } from '@/components/HorizontalLinearStepper/utils/getStep';
+import { StyledStepIcon } from '@/pages/UploadPage/components/StyledStepIcon';
 
-import { StyledStepIcon } from '../../pages/UploadPage/components/StyledStepIcon';
 import {
   StepperWrapper,
   StyledConnector,
