+++ src/pages/AnalysisResultPage/components/BloodTestSummary/index.tsx
@@ -1,14 +1,9 @@
 import { useRef } from 'react';
 
 import { MarkerTable } from '@/components/MarkerTable';
-import type { MarkerTableRef } from '@/components/MarkerTable';
 import { useMarkerTable } from '@/components/MarkerTable/useMarkerTable';
-
-import type { FullBloodTestAnalysisResult } from '../../types';
-
-interface BloodTestSummaryProps {
-  data?: FullBloodTestAnalysisResult;
-}
+import type { BloodTestSummaryProps } from '@/pages/AnalysisResultPage/types/types';
+import type { MarkerTableRef } from '@/types/marker';
 
 export const BloodTestSummary = ({ data }: BloodTestSummaryProps) => {
   const markerTableRef = useRef<MarkerTableRef>(null);
