+++ src/pages/AnalysisResultPage/components/HealthBar/index.tsx
@@ -1,12 +1,7 @@
-import { MARKER_STATUS_CLASSES } from '@/constants/marker';
+import { MARKER_STATUS_CLASSES } from '@/constants';
+import type { HealthBarProps } from '@/pages/AnalysisResultPage/types/types';
 
 import { PositionIndicator, ProgressBar, StretchedIcon } from './styles';
-
-interface HealthBarProps {
-  position: number;
-  status: string;
-  isSmallScreen: boolean;
-}
 
 const getStatusClass = (status: string): string => {
   return MARKER_STATUS_CLASSES[status as keyof typeof MARKER_STATUS_CLASSES] || '';
