+++ src/pages/AnalysisResultPage/components/MarkerCardsTable/index.tsx
@@ -5,7 +5,6 @@ import { Collapse } from '@mui/material';
 import { t } from 'i18next';
 import { ChevronDown, ChevronUp, Trash } from 'lucide-react';
 
-import type { MarkerProps } from '@/components/Marker';
 import {
   DeleteButton,
   ErrorText,
@@ -16,11 +15,12 @@ import {
   StyledTextField,
 } from '@/components/Marker/styles';
 import { useMarker } from '@/components/Marker/useMarker';
-import { MARKER_OPTIONS, MARKER_STATUS_CLASSES, UNIT_OPTIONS } from '@/constants/marker';
+import { MARKER_OPTIONS, MARKER_STATUS_CLASSES, UNIT_OPTIONS } from '@/constants';
 import { DeleteMarkerModal } from '@/pages/AnalysisResultPage/components/DeleteMarkerModal';
 import { ButtonContainer } from '@/pages/AnalysisResultPage/components/DeleteMarkerModal/styles';
 import { calculateMarkerPosition } from '@/pages/AnalysisResultPage/utils/markerPositionCalculator';
 import { ContinueButton } from '@/pages/ReviewCustomizePage/styles';
+import type { MarkerProps } from '@/types/marker';
 
 import { HealthBar } from '../HealthBar';
 import { BackButton } from '../ResultButtons/styles';
