+++ src/pages/AnalysisResultPage/constants/index.ts
@@ -1,0 +1,4 @@
+export const WELLNESS_SCORE_THRESHOLDS = {
+  EXCELLENT: 85,
+  GOOD: 65,
+} as const;
