+++ src/pages/HomePage/index.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next';
 
-import { PageLayout } from '@/components/Layout';
+import { PageLayout } from '@/utils/Layout';
 
 import { About, FAQ, HowItWorks, MainContent, PrivacySection } from './export';
