+++ src/pages/LegalPages/TermsPage/index.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import ReactMarkdown from 'react-markdown';
 
-import { PageLayout } from '@/components/Layout';
 import { LegalContent } from '@/pages/LegalPages/styles';
+import { PageLayout } from '@/utils/Layout';
 
 import termsOfUseText from '../content/legal/terms-of-use.en.md?raw';
