+++ src/pages/LoginPage/components/LoginForm/useLoginForm.ts
@@ -1,12 +1,8 @@
 import { useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 
+import type { LoginFormInputs } from '@/pages/LoginPage/types';
 import { useSendMagicLinkMutation } from '@/store/authApi';
-
-export type LoginFormInputs = {
-  email: string;
-  message: string;
-};
 
 export const useLoginForm = () => {
   const {
