+++ src/pages/UploadPage/constants/index.ts
@@ -1,0 +1,7 @@
+export const UPLOAD_STATUS = {
+  Idle: 'idle',
+  Uploading: 'uploading',
+  Success: 'success',
+  Error: 'error',
+  Rejected: 'rejected',
+} as const;
