+++ src/pages/UploadPage/types/index.ts
@@ -1,0 +1,37 @@
+import type { DropzoneInputProps, DropzoneRootProps } from 'react-dropzone';
+
+import type { UPLOAD_STATUS } from '../constants';
+
+export interface StatusDisplay {
+  text: string;
+  color: string;
+}
+
+export interface DropzoneFileProps {
+  getRootProps: (props?: DropzoneRootProps) => DropzoneRootProps;
+  getInputProps: (props?: DropzoneInputProps) => DropzoneInputProps;
+  hasFiles: boolean;
+  isUploading: boolean;
+  files: File[];
+  uploadProgress: Record<string, number>;
+  uploadStatus: UploadStatus;
+  errorMessage: string;
+  statusDisplay: StatusDisplay;
+  handleRemoveFile: (fileName?: string, e?: React.MouseEvent) => void;
+}
+
+export interface HealthReportAnalyzerProps {
+  isLoading: boolean;
+  titleKey?: string;
+  descriptionKey?: string;
+}
+
+export interface UploadErrorStateProps {
+  onRetry: () => void;
+}
+
+export interface UploadStepProps {
+  onContinue: () => void;
+}
+
+export type UploadStatus = (typeof UPLOAD_STATUS)[keyof typeof UPLOAD_STATUS];
